esra-explain-controller/requirements.txt
Flask==2.2.2
pymongo==4.3.3
python-dotenv==0.21.0
gevent==22.10.2